enhancement(listAuthors): tweak slug param
The `name` and `slug` params now accept a comma separated list of
values.  Before, multiple values had to be in a pipe separated list (in
an effort to be consistent with the behavior of other filter) However,
I think a comma separated list is more intuitive for this param, so
it now accepts both.